user specific - JWT

+++ backend/controllers/expense.js
@@ -7,7 +7,7 @@ const addexpense = (req, res) => {
     return res.status(400).json({ success: false, err: "Parameters missing" });
   }
 
-  Expense.create({ expenseamount, description, category })
+  Expense.create({ expenseamount, description, category, userId: req.user.id })
     .then((expense) => {
       return res.status(201).json({ expense, success: true });
     })
@@ -17,7 +17,7 @@ const addexpense = (req, res) => {
 };
 
 const getexpenses = (req, res) => {
-  Expense.findAll()
+  Expense.findAll({ where: { userId: req.user.id } })
     .then((expenses) => {
       return res.status(200).json({ expenses, success: true });
     })
@@ -32,7 +32,7 @@ const deleteexpense = (req, res) => {
   if (expenseid == undefined || expenseid.length === 0) {
     return res.status(400).json({ success: false });
   }
-  Expense.destroy({ where: { id: expenseid } })
+  Expense.destroy({ where: { id: expenseid, userId: req.user.id } })
     .then((noofrows) => {
       if (noofrows === 0) {
         return res.status(404).json({
